ZZ-885
1) Added share icons to page header when theme is 'ebooks-expanded'.
2) Don't add search if theme is 'ebooks-expanded'
